@@ -1,17 +1,17 @@
-sap.ui.define(['sap/fe/test/ObjectPage'], function(ObjectPage) {
-    'use strict';
+sap.ui.define(["sap/fe/test/ObjectPage"], function (ObjectPage) {
+  "use strict";
 
-    var CustomPageDefinitions = {
-        actions: {},
-        assertions: {}
-    };
+  const CustomPageDefinitions = {
+    actions: {},
+    assertions: {},
+  };
 
-    return new ObjectPage(
-        {
-            appId: 'sap.fe.cap.travel-analytics',
-            componentId: 'BookingsObjectPage',
-            entitySet: 'Bookings'
-        },
-        CustomPageDefinitions
-    );
+  return new ObjectPage(
+    {
+      appId: "sap.fe.cap.travel_analytics",
+      componentId: "BookingsObjectPage",
+      entitySet: "Bookings",
+    },
+    CustomPageDefinitions
+  );
 });
