@@ -1,4 +1,4 @@
-sap.ui.define(
+sap.ui.require(
   [
     "sap/fe/test/JourneyRunner",
     "sap/fe/cap/travel/test/integration/pages/MainListReport",
@@ -6,21 +6,27 @@ sap.ui.define(
     "sap/fe/cap/travel/test/integration/pages/ItemObjectPage",
     "sap/fe/cap/travel/test/integration/OpaJourney",
   ],
-  function (JourneyRunner, MainListReport, MainObjectPage, ItemObjectPage, Journey) {
+  function (
+    JourneyRunner,
+    MainListReport,
+    MainObjectPage,
+    ItemObjectPage,
+    Journey
+  ) {
     "use strict";
 
-    var journeyRunner = new JourneyRunner({
+    const runner = new JourneyRunner({
       // start index.html in web folder
       launchUrl: sap.ui.require.toUrl("sap/fe/cap/travel") + "/index.html",
       opaConfig: { timeout: 30 },
     });
 
-    journeyRunner.run(
+    runner.run(
       {
         pages: {
           onTheMainPage: MainListReport,
           onTheDetailPage: MainObjectPage,
-          onTheDetailItemPage: ItemObjectPage
+          onTheDetailItemPage: ItemObjectPage,
         },
       },
       Journey.run
