@@ -2,7 +2,7 @@ sap.ui.define(["sap/fe/test/ObjectPage"], function (ObjectPage) {
   "use strict";
 
   // OPTIONAL
-  var AdditionalCustomObjectPageDefinition = {
+  const AdditionalCustomObjectPageDefinition = {
     actions: {},
     assertions: {},
   };
