@@ -2,8 +2,11 @@
 window.suite = function () {
   "use strict";
 
-  let oSuite = new parent.jsUnitTestSuite();
-  let sContextPath = location.pathname.substring(0, location.pathname.lastIndexOf("/"));
+  const oSuite = new parent.jsUnitTestSuite();
+  const sContextPath = location.pathname.substring(
+    0,
+    location.pathname.lastIndexOf("/")
+  );
 
   oSuite.addTestPage(`${sContextPath}/integration/Opa.qunit.html`);
 
