@@ -38,28 +38,28 @@ public class AcceptRejectHandler implements EventHandler {
 	@Before(entity = Travel_.CDS_NAME)
 	public void beforeAcceptTravel(final AcceptTravelContext context) {
 
-		draftService.run(context.getCqn()).first(Travel.class).ifPresent(this::checkIfTravelHasExceptedStatus);
+		draftService.run(context.cqn()).first(Travel.class).ifPresent(this::checkIfTravelHasExceptedStatus);
 	}
 
 	@Before(entity = Travel_.CDS_NAME)
 	public void beforeRejectTravel(final RejectTravelContext context) {
-		draftService.run(context.getCqn()).first(Travel.class).ifPresent(this::checkIfTravelHasExceptedStatus);
+		draftService.run(context.cqn()).first(Travel.class).ifPresent(this::checkIfTravelHasExceptedStatus);
 	}
 
 	@On(entity = Travel_.CDS_NAME)
 	public void onRejectTravel(final RejectTravelContext context) {
-		Travel travel = draftService.run(context.getCqn()).single(Travel.class);
+		Travel travel = draftService.run(context.cqn()).single(Travel.class);
 		context.getCdsRuntime().requestContext().privilegedUser().run(ctx -> {
-			updateStatusForTravelId(travel.getTravelUUID(), TRAVEL_STATUS_CANCELLED, travel.getIsActiveEntity());
+			updateStatusForTravelId(travel.travelUUID(), TRAVEL_STATUS_CANCELLED, travel.isActiveEntity());
 		});
 		context.setCompleted();
 	}
 
 	@On(entity = Travel_.CDS_NAME)
 	public void onAcceptTravel(final AcceptTravelContext context) {
-		Travel travel = draftService.run(context.getCqn()).single(Travel.class);
+		Travel travel = draftService.run(context.cqn()).single(Travel.class);
 		context.getCdsRuntime().requestContext().privilegedUser().run(ctx -> {
-			updateStatusForTravelId(travel.getTravelUUID(), TRAVEL_STATUS_ACCEPTED, travel.getIsActiveEntity());
+			updateStatusForTravelId(travel.travelUUID(), TRAVEL_STATUS_ACCEPTED, travel.isActiveEntity());
 		});
 		context.setCompleted();
 	}
@@ -78,10 +78,10 @@ public class AcceptRejectHandler implements EventHandler {
 	}
 
 	private void checkIfTravelHasExceptedStatus(Travel travel) {
-		if (travel.getTravelStatusCode() != null && !travel.getTravelStatusCode()
+		if (travel.travelStatusCode() != null && !travel.travelStatusCode()
 				.equalsIgnoreCase(AcceptRejectHandler.TRAVEL_STATUS_OPEN)) {
-			throw new IllegalTravelStatusException("error.travel.status.unexpected", travel.getTravelID(),
-					AcceptRejectHandler.TRAVEL_STATUS_OPEN, travel.getTravelStatusCode());
+			throw new IllegalTravelStatusException("error.travel.status.unexpected", travel.travelID(),
+					AcceptRejectHandler.TRAVEL_STATUS_OPEN, travel.travelStatusCode());
 		}
 	}
 }
